ZoomCamp/Week2/Azure Airflow Ingestion/requirements.txt
pandas
pyarrow
fastparquet
azure-storage-blob